kaganNicholasIteration1
I was able to add a clear line of sight for the enemies. I also added a new starting point which eventually leads to the original starting point from earlier builds. This serves as an additional level to the game. I also reworked and added more obstacles to towards the ending and in a way making it its own level, with a platforming challenge that is more difficult than previous challenges in the game to end it off.